\name{hzar-package}
\alias{hzar-package}
\alias{hzar}
\docType{package}
\title{
  Hybrid Zone Analysis using R
}
\description{
  A collection of tools for modeling the shape of 1 dimensional clines.
}
\details{
\tabular{ll}{
Package: \tab hzar\cr
Type: \tab Package\cr
Version: \tab 0.1-7\cr
Date: \tab 2012-03-29\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
%~~ An overview of how to use the package, including the most important ~~
%~~ functions ~~
}
\author{
 Graham Derryberry
Maintainer:  Graham Derryberry <asterion@alum.mit.edu>
%~~ The author and/or maintainer of the package ~~
}
\references{
%~~ Literature or other references for background information ~~
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
\keyword{ package }
\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%~~ simple examples of the most important functions ~~

}
